requirements.txt
absl-py==0.9.0
alpaca-trade-api==0.42
apt-clone==0.2.1
apturl==0.5.2
astor==0.8.1
asyncio-nats-client==0.10.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.0
Brlapi==0.6.6
cachetools==4.0.0
certifi==2018.1.18
chardet==3.0.4
Click==7.0
command-not-found==0.3
configobj==5.0.6
cupshelpers==1.0
cycler==0.10.0
dash==1.8.0
dash-core-components==1.7.0
dash-html-components==1.0.2
dash-renderer==1.2.3
dash-table==4.6.0
decorator==4.4.1
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
entrypoints==0.3
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.1
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.9.2
idna==2.6
importlib-metadata==1.4.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.0
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.0
matplotlib==3.1.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
nemo-emblems==4.4.0
netifaces==0.10.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
onboard==1.4.1
opt-einsum==3.1.0
PAM==0.4.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.0
patsy==0.5.1
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
plotly==4.5.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.4.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.5.2
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.6
pyparted==3.11.1
pyRFC3339==1.0
pyrsistent==0.15.7
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-xapp==1.8.1
python-xlib==0.20
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.1
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
retrying==1.3.3
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
screen-resolution-extra==0.0.0
seaborn==0.10.0
Send2Trash==1.5.0
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.14.0
sklearn==0.0
statsmodels==0.11.1
system-service==0.3
systemd-python==234
tensorboard==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tinycss==0.4
tornado==6.0.3
traitlets==4.3.3
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.22
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.0
wrapt==1.11.2
xkit==0.0.0
zipp==2.1.0
